Nightmare mystery WTF?
So hit a major performance problem while filling cells.  With areas of
~400, we'd hang the editor for 2+ seconds.

Dug, and dug and dug until I boiled the problem down to the nonsense
example in `HexMap::_fill_cells()`, and I still don't have a solution.

Basically when I call `set_cell_item()` we get a huge delay while
calling the function.  Trick is, I stripped **everything** out of the
function; it's a noop.  Just returns.

So I created another function `do_nothing()`.  It's exactly the same,
just an empty function.  **The problem does not reproduce** calling
`do_nothing()`.

The assembly of the two functions is exactly the same.

The only time I've seen the problem go away is when I copied the
definition of `set_cell_item()` and renamed it to `do_nothing()`.  When
I compiled the problem went away, but as soon as I changed the order of
operations in `_fill_cells()`, the problem reproduced.

I've tried:
* clean build; no change
* renaming `set_cell_item()`; works, but I don't understand why

I'm saving everything here, and the disassembly in a gist[^1] to
document what the insanity this is.

[^1]: https://gist.github.com/dmlary/e70754d5386358b36f433a443daf69d5